feat(useCardRegister): 카드 정보 상태관리하는 훅 만들어서 코드 간소화
