+++ src/docs/README.md
@@ -51,3 +51,5 @@
 카드번호 타입이랑 실제 카드 브랜드가 일치 안하면 에러
 
 - 이름 입력시 영어 대문자만 입력이 되게 한다.
+- 마지막에 이름 잘 입력시 확인 버튼 안 생기는 에러
+- 최종적으로 내보내는 정보 객체 형태로 반환
