+++ src/pages/CardRegisterPage/CardRegisterPage.tsx
@@ -1,82 +1,23 @@
 import S from "./style";
 import CardRegisterForm from "@/components/CardRegisterForm/CardRegisterForm";
-import { MAX_LENGTH, VALID_LENGTH } from "@/constants/condition";
-import useInput from "@/hooks/useInput";
-import useInputs from "@/hooks/useInputs";
-import {
-  validateDoubleSpace,
-  validateEnterRequired,
-  validateIsValidLength,
-  validateMonth,
-  validateOwnerName,
-} from "@/utils/validation";
-import { CardNumberInputType } from "@/components/CardRegisterForm/components/CardNumbersField/CardNumbersField";
-import { ExpirationPeriodInputType } from "@/components/CardRegisterForm/components/ExpirationPeriodField/ExpirationPeriodField";
-import { CardType } from "@/constants/cardType";
 import CardPreview from "@/components/CreditCardPreview/CardPreview";
 import { useState } from "react";
 import BasicButton from "@/components/_common/BasicButton/BasicButton";
 import { theme } from "@/style/theme";
 import { useNavigate } from "react-router-dom";
+import useCardRegister from "@/hooks/useCardRegister";
 
 const CardRegisterPage = () => {
   const navigate = useNavigate();
-  const cardNumbersState = useInputs<CardNumberInputType>({
-    initialValue: {
-      cardNumbers1: "",
-      cardNumbers2: "",
-      cardNumbers3: "",
-      cardNumbers4: "",
-    },
-    validates: [
-      (value: string[]) =>
-        validateIsValidLength(value[0], VALID_LENGTH.CARD_NUMBERS),
-    ],
-    maxNumberLength: MAX_LENGTH.CARD_NUMBERS,
-  });
-
-  const expiredDateState = useInputs<ExpirationPeriodInputType>({
-    initialValue: { expirationMonth: "", expirationYear: "" },
-    validates: [
-      (value: string[]) =>
-        validateIsValidLength(value[0], VALID_LENGTH.EXPIRATION_PERIOD),
-      (value: string[], name: string) => {
-        if (name === "expirationMonth") {
-          return validateMonth(Number(value));
-        }
-        return null;
-      },
-    ],
-    maxNumberLength: MAX_LENGTH.EXPIRATION_PERIOD,
-  });
-
-  const ownerNameState = useInput({
-    initialValue: "",
-    validates: [
-      (value: string) => validateOwnerName(value),
-      (value: string) => validateDoubleSpace(value),
-      () => validateEnterRequired(),
-    ],
-  });
-
-  const cardTypeState = useInput<CardType | null>({
-    initialValue: null,
-    validates: [(value: string) => validateOwnerName(value)],
-  });
-
-  const CVCNumbersState = useInput<string>({
-    initialValue: "",
-    validates: [
-      (value: string) => validateIsValidLength(value, VALID_LENGTH.CVC_NUMBERS),
-    ],
-  });
-
-  const passwordState = useInput<string>({
-    initialValue: "",
-    validates: [
-      (value: string) => validateIsValidLength(value, VALID_LENGTH.PASSWORD),
-    ],
-  });
+  const cardRegister = useCardRegister();
+  const {
+    cardNumbersState,
+    expirationPeriodState,
+    cardTypeState,
+    CVCNumbersState,
+    passwordState,
+    ownerNameState,
+  } = cardRegister;
 
   const [step, setStep] = useState<number>(1);
 
@@ -92,7 +33,7 @@ const CardRegisterPage = () => {
   const completedArr = [
     !cardNumbersState.isError,
     !!cardTypeState.value?.length,
-    !expiredDateState.isError,
+    !expirationPeriodState.isError,
     !ownerNameState.isError,
     !CVCNumbersState.isError,
     !passwordState.isError,
@@ -106,18 +47,13 @@ const CardRegisterPage = () => {
         <CardPreview
           cardType={cardTypeState.value}
           cardNumbers={cardNumbersState.values}
-          expirationDate={expiredDateState.values}
+          expirationDate={expirationPeriodState.values}
           ownerName={ownerNameState.value}
           CVCNumbers={CVCNumbersState.value}
           step={step}
         />
         <CardRegisterForm
-          cardNumbersState={cardNumbersState}
-          expiredPeriodState={expiredDateState}
-          ownerNameState={ownerNameState}
-          cardTypeState={cardTypeState}
-          CVCNumbersState={CVCNumbersState}
-          passwordState={passwordState}
+          {...cardRegister}
           step={step}
           setStep={setStep}
           completedArr={completedArr}
